test: CuratorsPicksEmerging
